Decommission old hours library
